cpuinfo_serial: Fix error when sysfs serial_number is missing
We need to not print errors when doing the processing and look up of the
information, as UpdateHub uses the output as a interface. So we now send
errors to /dev/null.

While reworking it, we also changed the code flow to avoid calling the
code branches as soon as we get the first result.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>